% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/density_grid_stats.R
\name{get_p_dir}
\alias{get_p_dir}
\title{Compute Probability of Direction from Discretised Density}
\usage{
get_p_dir(
  val,
  dens,
  alternative = c("two.sided", "greater", "less"),
  dx = NULL
)
}
\arguments{
\item{val}{Numeric vector representing the grid of values over which the
density is defined.}

\item{dens}{Numeric vector of the same length as \code{val}, representing the
density at each grid point.}

\item{alternative}{Character string specifying the alternative hypothesis.
One of \code{"two.sided"} (default), \code{"greater"}, \code{"less"}.}

\item{dx}{Grid spacing (step size). If \code{NULL} (default), it is estimated
from the unique values of \code{val} (grid points) using \code{\link[=diff]{diff()}}.}
}
\value{
A numeric value corresponding to the probability of direction.
}
\description{
Internal helper function to compute the probability of direction
from a numeric vector of values and a corresponding density over a
discretised grid.
}
\details{
The probability of direction (PD) represents the proportion of the
distribution that lies on one side of zero. For example, a PD of \code{0.92} means
that 92\% of the distribution's mass lies one one side (strictly positive or
strictly negative values, whichever is greatest). In the context of
hypothesis testing, it can be taken as an index of the \emph{existence} of an
effect, but does \emph{not} necessarily address the size, importance, or
\emph{"significance"} of an effect. For example, a vanishingly small effect whose
distribution is precisely concentrated within the
\eqn{\left[0.0001, 0.0002\right]} range would still have a PD of
approximately \code{1}.

Note that the notion of the probability of direction is undefined for one-sided
hypotheses. If \verb{alternative != "two.sided}, the returned value is \code{NA_real_}.
}
\keyword{internal}
