% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/density_grid_stats.R
\name{get_p_rope}
\alias{get_p_rope}
\title{Perform ROPE-based Equivalence Test from Discretised Density}
\usage{
get_p_rope(
  val,
  dens,
  rope_range,
  ci_range = NULL,
  alternative = c("two.sided", "greater", "less"),
  dx = NULL
)
}
\arguments{
\item{val}{Numeric vector representing the grid of values over which the
density is defined.}

\item{dens}{Numeric vector of the same length as \code{val}, representing the
density at each grid point.}

\item{rope_range}{Numeric vector of length 2 specifying the lower and upper
bounds of the ROPE.}

\item{ci_range}{Optional numeric vector of length 2 specifying the lower and
bounds of a credible interval (CI) of the distribution. If \code{NULL}
(default), the returned value represents the proportion of the \emph{full distribution}
contained within the ROPE. If a CI of the distribution is provided,
the returned value represents the proportion of the \emph{distribution's CI}
contained within the ROPE.}

\item{alternative}{Character string specifying the alternative hypothesis.
One of \code{"two.sided"} (default), \code{"greater"}, \code{"less"}.}

\item{dx}{Grid spacing (step size). If \code{NULL} (default), it is estimated
from the unique values of \code{val} (grid points) using \code{\link[=diff]{diff()}}.}
}
\value{
A numeric value corresponding to the proportion of the distribution
(or credible interval thereof) contained within the ROPE.
}
\description{
Internal helper function to perform an equivalence test, based
on a numeric vector of values and a corresponding density over a discretised
grid. In equivalence testing, the null hypothesis is that a parameter value
falls within a specified region of practical equivalence (ROPE). The
proportion of the distribution (or a credible interval thereof) contained
within the ROPE can be used as a decision rule regarding the "significance"
of a parameter.
}
\details{
The ROPE-based p-value allows for a pragmatic interpretation of the
"significance" of an effect, indicating if a parameter is effectively zero
for all practical purposes. It is commonly computed as the proportion of the
distribution's 95\% CI contained within the ROPE (Kruschke & Liddell, 2018).
If \code{p_rope} is approximately equal to 1, (almost) all of the distribution's
most-credible values are inside the ROPE, suggesting that the null hypothesis
can be accepted. If \code{p_rope} is approximately equal to 0, (almost) all of
the distribution's most-credible values are outside the ROPE, suggesting
that the null hypothesis can be rejected. Otherwise, it remains unclear if
the null hypothesis should be accepted or rejected.
However, some have argued that computing the proportion of the \emph{full distribution}
(as opposed to the \emph{distribution's CI}) contained within the ROPE provides a
more sensitive test (Makowski et al., 2019). In this case, \code{p_rope} values
less than 0.025 and greater than 0.975 are used as thresholds to reject or
accept the null hypothesis, respectively.
}
\references{
Makowski, D., Ben-Shachar, M. S., Chen, S. H. A., & Lüdecke, D. (2019).
Indices of effect existence and significance in the Bayesian framework.
\emph{Frontiers in Psychology}, \emph{10}. \doi{10.3389/fpsyg.2019.02767}

Kruschke, J. K., Liddell, T. M. (2018). The Bayesian new statistics:
Hypothesis testing, estimation, meta-analysis, and power analysis from a
Bayesian perspective. \emph{Psychonomic Bulletin & Review}, \emph{25}(1), 178-206.
\doi{10.3758/s13423-016-1221-4}
}
\keyword{internal}
