% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plausible_cor.R
\name{get_posterior_rho_densities}
\alias{get_posterior_rho_densities}
\title{Evaluate posterior density functions over a grid}
\usage{
get_posterior_rho_densities(.data, grid_spacing = 0.001)
}
\arguments{
\item{.data}{Data frame output from [run_plausible_cor()].}

\item{grid_spacing}{The step size for the grid of correlation values to
evaluate.}
}
\value{
A data frame where each row corresponds to one density value for a
        given posterior density function.
}
\description{
Helper function that evaluates each posterior density function in the output
of [run_plausible_cor()] over a shared grid from -1 to 1. Typical users
would not have to call this function, instead relying on
[summarise_plausible_cor()], [compare_plausible_cors()], and
[plot_population_cor()].
}
