% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/posterior.R
\name{likelihood_rho_exact}
\alias{likelihood_rho_exact}
\alias{likelihood_rho_jeffreys}
\title{Likelihood of the sample Pearson's correlation}
\usage{
likelihood_rho_exact(r, n, cor_grid, max_iter)

likelihood_rho_jeffreys(r, n, cor_grid)
}
\arguments{
\item{r}{Numeric value. The observed sample correlation coefficient.}

\item{n}{Numeric value. The sample size.}

\item{cor_grid}{Numeric vector. Grid of correlation values to evaluate.}

\item{max_iter}{Integer. Maximum number of iterations (attempts) to solve
generalised hypergeometric functions.}
}
\value{
Numeric vector of likelihood values.
}
\description{
Evaluates the exact sampling distribution of the sample correlation
coefficient \eqn{r} given the population correlation \eqn{\rho}.
This is expressed in terms of generalized hypergeometric functions. Instead
of the exact solution, Jeffrey's approximation can be used, which is accurate
when the sample correlation is (practically) equal to zero.
}
\keyword{internal}
