% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{plot_population_cor}
\alias{plot_population_cor}
\title{Plot correlation posterior distributions from plausible value analysis}
\usage{
plot_population_cor(
  .data,
  n_draws = 500,
  trace_aes = NULL,
  mean_aes = NULL,
  mean_interval_args = NULL,
  mean_interval_aes = NULL,
  sample_rug_aes = NULL,
  zero_refline_aes = NULL,
  x_title = "population-level correlation",
  x_axis_limits = NULL,
  plot_text_scaling = 1,
  rng_seed = NULL
)
}
\arguments{
\item{.data}{A data frame returned by \code{\link[=run_plausible_cor]{run_plausible_cor()}}, containing at
least \code{.draw}, \code{r}, and \code{posterior_updf} columns.}

\item{n_draws}{Integer specifying the maximum number of MCMC draws to plot.
Defaults to \code{500} to avoid overplotting of density traces.}

\item{trace_aes}{A named list of aesthetics for the individual posterior
density traces (e.g., \code{colour}, \code{alpha}, \code{linewidth}), passed to
\code{\link[ggplot2:geom_path]{ggplot2::geom_line()}}. Defaults to
\code{list(colour = "#4C4C4C", alpha = 0.1, linewidth = 0.1)}. Set to
\code{FALSE} to omit these lines entirely.}

\item{mean_aes}{A named list of aesthetics for the mean posterior density
trace (e.g., \code{colour}, \code{alpha}, \code{linewidth}), passed to
\code{\link[ggplot2:geom_path]{ggplot2::geom_line()}}. Defaults to
\code{list(colour = "#377EB8", linewidth = 1.5)}. Set to
\code{FALSE} to omit the mean trace entirely.}

\item{mean_interval_args}{A named list that specifies how an interval for the
mean posterior density is computed. Should include the elements \code{width},
which is a single numeric value of the desired interval width, and \code{method},
which is a character specifying which interval type to compute.
Defaults to \code{list(width = 0.95, method = "hdci")}.}

\item{mean_interval_aes}{A named list of aesthetics for a light shaded
background rectangle that represents the requested interval of the
mean posterior density trace. Arguments are passed to \code{\link[ggplot2:annotate]{ggplot2::annotate()}}.
Defaults to \code{list(fill = "#BDD7E7", alpha = 0.4)}.
Set to \code{FALSE} to omit the interval entirely.}

\item{sample_rug_aes}{A named list of aesthetics for "rug" plot of the sample
correlation coefficients on which the individual posterior density
traces are based. Arguments are passed to \code{\link[ggplot2:geom_rug]{ggplot2::geom_rug()}}.
Defaults to \code{list(alpha = 0.1, linewidth = 0.1, length = grid::unit(0.03, "npc"))}.
Set to \code{FALSE} to omit the rug plot entirely.}

\item{zero_refline_aes}{A named list of aesthetics for the vertical reference
line at zero (e.g., \code{linetype}, \code{linewidth}, \code{colour}), passed to
\code{\link[ggplot2:geom_abline]{ggplot2::geom_vline()}}. Defaults to
\code{list(linetype = "dashed", linewidth = 0.75)}. Set to \code{FALSE} to omit
the reference line entirely.}

\item{x_title}{Label for the x-axis. Defaults to \code{"population-level correlation"}}

\item{x_axis_limits}{Numeric vector of length 2 that defines x-axis limits
to "zoom" into, using \code{\link[ggplot2:coord_cartesian]{ggplot2::coord_cartesian()}}. Alternatively,
\code{NULL} (default) retains the full scale.}

\item{plot_text_scaling}{Numeric scaling factor for axis text and title
size. Default is \code{1}.}

\item{rng_seed}{Optional integer seed to ensure reproducible sampling of
posterior draws.}
}
\value{
A \link[ggplot2:ggplot]{ggplot2::ggplot} object.
}
\description{
Visualises density traces of posterior distributions of Pearson correlation
coefficients obtained from \code{\link[=run_plausible_cor]{run_plausible_cor()}}, along with the mean
posterior density across all MCMC samples.
}
\details{
This function plots the plausible posterior distributions of the Pearson
correlation coefficient obtained via \code{\link[=run_plausible_cor]{run_plausible_cor()}}. It displays:
\itemize{
\item Density traces for a subset (\code{n_draws}) of individual posterior draws,
capturing both parameter uncertainty and population variability.
\item The mean posterior density across all draws, offering a summary of the
plausible population-level correlation.
\item An optional vertical reference line at 0 to facilitate visual
interpretation.
}

Subsampling with \code{n_draws} is used to limit visual clutter when many MCMC
draws are available. However, the mean posterior density is always computed
from the full set of draws.

Each plotted layer can be customized using a named list of aesthetic values,
or disabled entirely by passing \code{FALSE}.
}
\examples{
# Typical usage ------------------------------------------------------------
# for demonstration purposes only, run with small subset of MCMC samples
caution_striatum_cor <- run_plausible_cor(
  parameter = "caution_effect_speed",
  covariate = "striatum",
  mcmc_data = Forstmann_LBA,
  covariate_data = Forstmann_fMRI,
  n_draws = 500,
  rng_seed = 123
)
plot_population_cor(caution_striatum_cor)
# has equivalent S3 generic
\dontrun{
  plot(caution_striatum_cor)
}
# can be used with magrittr pipe operator
\dontrun{
  library(magrittr)
  caution_striatum_cor \%>\%
    plot_population_cor()
}

# Non-default plot settings ------------------------------------------------
# simplify plot by setting certain plot elements to FALSE
\dontrun{
  plot_population_cor(
    caution_striatum_cor,
    trace_aes = FALSE,
    mean_interval_aes = FALSE,
    sample_rug_aes = FALSE,
    zero_refline_aes = FALSE
  )
}
# change settings of specific plot element
\dontrun{
  plot_population_cor(
    caution_striatum_cor,
    mean_interval_aes = list(
      colour = "red", linewidth = 3
    )
  )
}
# zoom in on specific x-axis range
\dontrun{
  plot_population_cor(
    caution_striatum_cor,
    x_axis_limits = c(NA, 0.5)
  )
}

}
\seealso{
\code{\link[=run_plausible_cor]{run_plausible_cor()}}
}
