% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S3_generics.r
\name{print.plausible_cor_list}
\alias{print.plausible_cor_list}
\title{Print method for plausible correlation list objects}
\usage{
\method{print}{plausible_cor_list}(x, ...)
}
\arguments{
\item{x}{An object of class "plausible_cor_list" (a list of plausible_cor objects).}

\item{...}{Optional additional arguments (ignored).}
}
\description{
Print method for plausible correlation list objects
}
