% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S3_generics.r
\name{print.summary.plausible_cor}
\alias{print.summary.plausible_cor}
\title{Print method for summary.plausible_cor objects}
\usage{
\method{print}{summary.plausible_cor}(x, ...)
}
\arguments{
\item{x}{An object of class "summary.plausible_cor".}

\item{...}{Optional additional arguments, passed to tibble's print method.}
}
\description{
Print method for summary.plausible_cor objects
}
