% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S3_generics.r
\name{print.summary.plausible_cor_list}
\alias{print.summary.plausible_cor_list}
\title{Print method for summary.plausible_cor_list objects}
\usage{
\method{print}{summary.plausible_cor_list}(x, ...)
}
\arguments{
\item{x}{An object of class "summary.plausible_cor_list".}

\item{...}{Additional arguments passed to tibble's print method.}
}
\description{
Print method for summary.plausible_cor_list objects
}
