% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plausible_cor.R
\name{summarise_samples}
\alias{summarise_samples}
\title{Summarise posterior samples}
\usage{
summarise_samples(
  .data,
  ...,
  point_method = c("mean", "median"),
  interval_width = 0.95,
  interval_method = c("hdci", "qi"),
  rope_range = NULL
)
}
\arguments{
\item{.data}{A data frame containing a single numeric column representing
posterior samples or any distribution of interest.}

\item{...}{Additional arguments passed to \code{\link[ggdist:point_interval]{ggdist::point_interval()}}.}

\item{point_method}{A string indicating the point estimate to compute. Can
either be \code{"mean"} (default) or \code{"median"}.}

\item{interval_width}{Numeric vector of values between 0 and 1 specifying the
width(s) of the credible interval(s) to calculate. Default is \code{0.95}.}

\item{interval_method}{A string indicating the method for computing intervals.
Can either be \code{"hdci"} (highest-density continous intervals; default) or
\code{"qi"} (quantile intervals).}

\item{rope_range}{Optional numeric vector of length 2 specifying the lower
and upper bounds of the region of practical equivalence (ROPE), which
is used to compute the proportion of the distribution contained within
the ROPE. Defaults to \code{NULL} in which case the ROPE is ignored.}
}
\value{
A \link[tibble:tibble]{tibble::tibble} summarizing the posterior samples.
}
\description{
Summarises a single numeric column from a data frame containing posterior
samples (or any numeric data), providing a point estimate (mean or median),
credible interval(s) (e.g., highest-density continous intervals or quantiles),
the directionality of the distribution (positive or negative), and optionally
the proportion contained in a region of practical equivalence (ROPE).
}
\keyword{internal}
