% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plausible_cor.R
\name{compare_plausible_cors}
\alias{compare_plausible_cors}
\title{Compare two distributions of plausible correlation values}
\usage{
compare_plausible_cors(
  x,
  y,
  point_interval_args = NULL,
  rope_range = NULL,
  n_samples = 1,
  rng_seed = NULL,
  posterior_grid_spacing = 0.001
)
}
\arguments{
\item{x, y}{Data frames returned by \code{\link[=run_plausible_cor]{run_plausible_cor()}}, each containing
the columns \code{.draw}, \code{r}, and \code{posterior_updf}.}

\item{point_interval_args}{A named list specifying how the central tendency
and credible interval(s) should be computed, or \code{NULL} to use defaults.
Valid elements are:
\itemize{
\item \code{point_method} character string, either \code{"mean"} (default) or \code{"median"}
for the point summary.
\item \code{interval_method} character string, either \code{"hdci"} (default) for highest
density credible interval or \code{"qi"} for quantile interval.
\item \code{interval_width} numeric vector with values between 0 and 1 specifying
the desired interval width(s). Default is \code{0.95} for the 95\% credible
interval.
Any invalid arguments will be ignored with a warning. Partial
specification is supported; unspecified elements will use defaults.
}}

\item{rope_range}{Optional numeric vector of length 2 specifying the lower
and upper bounds of the region of practical equivalence (ROPE), which
is used to compute the proportion of the distribution contained within
the ROPE. Defaults to \code{NULL} in which case the ROPE is ignored.}

\item{n_samples}{Integer indicating how many quantiles to draw per posterior
distribution when \code{comparison_method = "population"}. Default is \code{1}.}

\item{rng_seed}{Optional numeric vector of length 2 to control the random
seed for quantile sampling of \code{x} and \code{y}, respectively. If \code{NULL}
(default), random seeds will not be set.}

\item{posterior_grid_spacing}{For comparing the population-level plausible
correlations: The step size used to discretise each MCMC sample's
posterior density function into a grid of posterior density values,
before quantiles are drawn. Default is \code{1e-3}.}
}
\value{
A \code{\link[tibble:tibble]{tibble::tibble()}} containing summary statistics of the difference
between correlation coefficients, with the same format as \code{\link[=summarise_plausible_cor]{summarise_plausible_cor()}}.
}
\description{
Compares the posterior distributions of plausible correlation values, as
returned by two separate calls to \code{\link[=run_plausible_cor]{run_plausible_cor()}}. Uses two methods
methods for defining the distribution of the difference: a \emph{sample-based}
comparison, which subtracts the correlation values from matching MCMC
samples, and a \emph{population-based} comparison, which draws quantiles from
each posterior distribution (accounting for population uncertainty) before
computing the difference.
}
\details{
For the sample-based comparison, the function computes the difference in the
Pearson correlation coefficient for each \code{.draw} shared between the two data
frames.

For the population-based comparison, the function draws quantile value(s)
from the posterior distribution of each Pearson correlation (as defined by
\code{posterior_updf}) and computes the difference. This allows for comparison at
the population level rather than the specific sample of subjects.

This function was adapted from code previously released with the Dynamic
Models of Choice toolbox (Heathcote et al., 2019).
}
\references{
Ly, A., Boehm, U., Heathcote, A., Turner, B. M., Forstmann, B., Marsman, M.,
& Matzke, D. (2017). A flexible and efficient hierarchical Bayesian approach
to the exploration of individual differences in cognitive‐model‐based
neuroscience. In A. A. Moustafa (Ed.),  \emph{Computational Models of Brain and
Behavior} (pp. 467–479). Wiley. \doi{10.1002/9781119159193.ch34}

Heathcote, A., Lin, Y.S., Reynolds, A., Strickland, L., Gretton, M., &
Matzke, D. (2019). Dynamic models of choice. \emph{Behavior Research Methods},
\emph{51}, 961-985. \doi{10.3758/s13428-018-1067-y}
}
