% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/posterior_rho.R
\name{posterior_rho_updf}
\alias{posterior_rho_updf}
\title{Calculate Unnormalised Posterior Density for Pearson's Correlation
Coefficient}
\usage{
posterior_rho_updf(r, n, kappa = 1, n_bins = 1000, ...)
}
\arguments{
\item{r}{Numeric value. The observed sample correlation coefficient, must be
between -1 and 1.}

\item{n}{Numeric value. The sample size, must be at least 3.}

\item{kappa}{Numeric value. Parameter controlling the "concentration" of the
stretched beta prior on the correlation coefficient (see details
below). Default is 1, resulting in a uniform prior.}

\item{n_bins}{Integer. Number of grid points for the approximation, default
is 1000.}

\item{...}{Additional arguments passed to [stats::approxfun()].}
}
\value{
A function that evaluates the unnormalized posterior density at any
        correlation value between -1 and 1.
}
\description{
This function calculates the unnormalised posterior density for Pearson's
correlation coefficient based on methods introduced by Ly et al. (2018).
It returns an approximation function that can be evaluated at any correlation
value between -1 and 1.
}
\details{
The function implements the analytical posterior for Pearson's correlation
coefficient as described in Ly et al. (2018). For r = 0, Jeffreys'
approxmation is used; otherwise the exact solution given by Ly et al. (2018)
is used.

The prior on the correlation coefficient is a so-called stretched beta
distribution, which is a beta distribution with its shape parameters alpha
and beta set to alpha = beta = 1/kappa, and its minimum and maximum
stretched to -1 and 1, respectively. This creates a symmetric distribution
centered at zero and its domain stretched to cover the full range of the
correlation coefficient.

This function was adapted from code previously released with the Dynamic
Models of Choice toolbox (Heathcote et al., 2019).
}
\references{
Ly, A., Marsman, M., & Wagenmakers, E.-J. (2018). Analytic posteriors for
Pearson's correlation coefficient. *Statistica Neerlandica*, 72, 4–13.
https://doi.org/10.1111/stan.12111

Heathcote, A., Lin, Y.S., Reynolds, A., Strickland, L., Gretton, M., &
Matzke, D. (2019). Dynamic models of choice. *Behavior Research Methods*,
51, 961-985. https://doi.org/10.3758/s13428-018-1067-y
}
