% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S3_generics.r
\name{print.plausible_cor}
\alias{print.plausible_cor}
\title{Print method for plausible correlation objects}
\usage{
\method{print}{plausible_cor}(x, ...)
}
\arguments{
\item{x}{An object of class "plausible_cor" (output from \code{\link[=run_plausible_cor]{run_plausible_cor()}}).}

\item{...}{Optional additional arguments, passed to tibble's print method.}
}
\description{
Print method for plausible correlation objects
}
