% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plausible_cor.R
\name{run_plausible_cor}
\alias{run_plausible_cor}
\title{Calculate plausible correlation values between model parameter and covariate}
\usage{
run_plausible_cor(
  mcmc_data,
  covariate_data = NULL,
  draw_id,
  subject_id,
  parameter,
  covariate,
  confounders = NULL,
  posterior_args = NULL
)
}
\arguments{
\item{mcmc_data}{Data frame of MCMC samples in long format, where each row
represents a unique combination of MCMC draw, subject, and
parameter value.}

\item{covariate_data}{Optional data frame containing covariate data (if not
in \code{mcmc_data}). Should contain one row per participant.}

\item{draw_id}{String denoting the column with MCMC sample IDs in \code{mcmc_data}.}

\item{subject_id}{String denoting the column with subject IDs. If both \code{mcmc_data}
and \code{covariate_data} are provided, this column name should be consistent
across data frames.}

\item{parameter}{String denoting the column with the estimated model
parameter.}

\item{covariate}{String denoting the column with the observed covariate of interest.}

\item{confounders}{Optional character vector denoting the column(s) in \code{covariate_data}
that are confounding or "controlling" variables, that is, variables that are
associated with both \code{parameter} and \code{covariate}, and therefore contaminate
the direct association between \code{parameter} and \code{covariate}.}

\item{posterior_args}{Named list of arguments passed to
\code{\link[=posterior_rho_updf]{posterior_rho_updf()}}, which controls how the unnormalised posterior
density function is computed for each Pearson correlation coefficient.
The following arguments can be included:
\itemize{
\item \code{kappa} Numeric value controlling the "concentration" of the
stretched beta prior on the correlation coefficient. Default is \code{1},
resulting in a uniform prior.
\item \code{n_bins} Integer denoting the number of bins ("resolution") used
for the discretised grid of correlation values, for which the
unnormalised posterior density values are computed. Default is \code{1e3}.
\item \code{max_iter} Integer denoting the maximum number of iterations
(attempts) to obtain a posterior density function for a given MCMC
sample. Default is \code{1e7}.
\item \code{...} additional arguments passed forward to \code{\link[stats:approxfun]{stats::approxfun()}}.
}}
}
\value{
A \link[tibble:tbl_df-class]{tibble::tbl_df} with one row per MCMC sample containing:
\item{.draw}{The MCMC sample ID}
\item{r}{Pearson correlation coefficient}
\item{n}{Number of complete observations used for correlation calculation}
\item{posterior_updf}{Function object for evaluating the unnormalised
posterior density at any correlation value between -1 and 1}
}
\description{
Implements a Bayesian approach for exploring individual differences in a
model parameter that was estimated without reference to the observed
covariate of interest. This function computes the Pearson correlation
coefficient between subject-wise parameter values (estimated using MCMC
sampling methods) and an observed covariate (e.g., questionnaire) for each
MCMC sample, resulting in a distribution of "plausible values" for the
correlation coefficient. It additionally provides the posterior density
function for each correlation coefficient, to enable inference about the
latent correlation in the population (as opposed to the given sample of
subjects).
}
\details{
This function implements the approach described by Ly et al. (2017) for
exploring individual differences in cognitive-model-based neuroscience:
\enumerate{
\item First, it computes the Pearson correlation between an estimated model
parameters and observed covariate separately for each MCMC sample,
resulting in a set of plausible correlation values for the given sample
of subjects.
If confounding variables are specified, the Pearson \emph{partial} correlation
is computed, that is, the Pearson correlation between the model parameter
and covariate while accounting for the effects of confounding variables.
This step handles uncertainty in the estimation of individual subjects'
model parameters, and is suitable for inferences about the correlation
coefficient in the given sample of subjects. However, it does not account
for uncertainty in generalising from the sample to the population.
\item Second, it calculates the posterior density function for each (partial)
correlation using the analytical solution derived by Ly et al. (2018),
resulting in a set of plausible posterior distributions of the correlation.
This handles uncertainty in generalizing from the sample to the
population, and is therefore suitable for inferences about the
correlation coefficient in the population.
}

The function requires at least 3 valid observations per MCMC sample and
filters out any correlation values that are not finite or outside the valid
range of \eqn{\left[-1, 1\right]}.
}
\references{
Ly, A., Boehm, U., Heathcote, A., Turner, B. M., Forstmann, B., Marsman, M.,
& Matzke, D. (2017). A flexible and efficient hierarchical bayesian approach
to the exploration of individual differences in cognitive‐model‐based
neuroscience. \emph{Computational models of brain and behavior}, 467-479. DOI:
\href{https://doi.org/10.1002/9781119159193.ch34}{10.1002/9781119159193.ch34}

Ly, A., Marsman, M., & Wagenmakers, E.-J. (2018). Analytic posteriors for
Pearson's correlation coefficient. \emph{Statistica Neerlandica}, \emph{72}, 4–13. DOI:
\href{https://doi.org/10.1111/stan.12111}{10.1111/stan.12111}

Kucharský, S., Wagenmakers, E.-J., Van den Bergh, D., & Ly, A. (2023).
Analytic posterior distributions and Bayes Factor for Pearson Partial
Correlation. \emph{PsyArXiv Preprints}. DOI:
\href{https://doi.org/10.31234/osf.io/6muwy}{10.31234/osf.io/6muwy}
}
