% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plausible_cor.R
\name{summarise_plausible_cor}
\alias{summarise_plausible_cor}
\title{Summarise plausible correlation estimates}
\usage{
summarise_plausible_cor(
  .data,
  point_interval_args = NULL,
  rope_range = NULL,
  posterior_grid_spacing = 0.001
)
}
\arguments{
\item{.data}{A data frame output from \code{\link[=run_plausible_cor]{run_plausible_cor()}}.}

\item{point_interval_args}{A named list specifying how the central tendency
and credible interval(s) should be computed, or \code{NULL} to use defaults.
Valid elements are:
\itemize{
\item \code{point_method} character string, either \code{"mean"} (default) or \code{"median"}
for the point summary.
\item \code{interval_method} character string, either \code{"hdci"} (default) for highest
density credible interval or \code{"qi"} for quantile interval.
\item \code{interval_width} numeric vector with values between 0 and 1 specifying
the desired interval width(s). Default is \code{0.95} for the 95\% credible
interval.
Any invalid arguments will be ignored with a warning. Partial
specification is supported; unspecified elements will use defaults.
}}

\item{rope_range}{Optional numeric vector of length 2 specifying the lower
and upper bounds of the region of practical equivalence (ROPE), which
is used to compute the proportion of the distribution contained within
the ROPE. Defaults to \code{NULL} in which case the ROPE is ignored. See
Gignac & Szodorai (2016) for guidance on what constitutes a correlation
coefficient practically equivalent to null.}

\item{posterior_grid_spacing}{For summarising the population-level plausible
correlation: The step size used to discretise each MCMC sample's
posterior density function into a grid of posterior density values.
Default is \code{1e-3}.}
}
\value{
A \link[tibble:tbl_df-class]{tibble::tbl_df} with one row per summary type
("sample" and "posterior") and credible interval width, containing:
\item{type}{Whether the row corresponds to the sample-based or
population-based summary.}
\item{mean / median}{Point estimate of the correlation coefficient.}
\item{lower / upper}{Credible interval bounds for each \code{interval_width}.}
\item{width}{If the length of \code{interval_width} is greater than 1: The width of the credible interval.}
\item{p_dir}{Directional probability (proportion of mass > 0 or < 0, whichever is greater). Only applicable if \code{\link[=run_plausible_cor]{run_plausible_cor()}} was called with \code{alternative = "two.sided"}.}
\item{p_rope}{If specified: The proportion contained within the ROPE.}
}
\description{
Takes the output of \code{\link[=run_plausible_cor]{run_plausible_cor()}} and summarises the distribution of
plausible correlation values at both the sample level (across MCMC samples)
and the population level (via the mean posterior density across MCMC
samples).
}
\references{
Gignac, G. E., & Szodorai, E. T. (2016). Effect size guidelines for
individual differences researchers. \emph{Personality and individual differences},
\emph{102}, 74-78. \doi{10.1016/j.paid.2016.06.069}
}
